requirements.txt
alembic==1.8.1
mysqlclient==2.1.1
paho-mqtt==1.6.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
